stock-analyzer/backend/requirements/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9

# Data Analysis & Technical Indicators
pandas==2.1.4
numpy==1.24.3
TA-Lib==0.4.28
yfinance==0.2.28
scipy==1.11.4

# News & Sentiment Analysis
newsapi-python==0.2.7
textblob==0.17.1
nltk==3.8.1
vaderSentiment==3.3.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
schedule==1.2.0
redis==5.0.1
celery==5.3.4
httpx==0.25.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0